+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Prevent using `/` in group name.
 - Decide group name from filename for `group()` call without args.
 - `assert` returns input values.
+- `assert[_not]_equals` works for Tarantool's box.tuple.
 
 # 0.3.0